Provide ordering guarantees as per Akka
We now provide the same ordering guarantees per Akka where events are processed immediately after the effect, and not deferred async. Similarly, effects are processed inline. This could reduce throughput, but I doubt that this will be an issue in practice. If it is, then we can introduce a task pool keyed by entity id.