Introduces source and flow
Record source and flow is introduced so that we can better convey intent, rather than passing closures, or consuming the entity manager. Further more, the act of emitting an event will now not proceed with any other combinators if it fails. This is in line with Akka Persistence and its combinators.