Encryption for commit log storage
Provides the hooks to encrypt/decrypt records for the marshaler and updates the example to illustrate. Encrypting data at rest with the commit log is important and was initially imposed by working with large vendors at the edge. Streambed supports rich support for encryption so this does not burden a developer.

I have also made the file log module generic over CommitLog and, now, an associated SecretStore. Streambed has a secret store named "Confidant", and also provides an implementation for connecting with Hashicorp Vault. Very little effort was expended in becoming more generic.