Make the entity manager a task
Simplifies the entity manager by flattening it into an async function. Entity managers are not shared, so this approach seems more reasonable. Spawning the task is then left to the discretion of the caller, which then becomes more direct and comprehensible. We also get the slight optimisation of one less spawn and its associated allocations.