Re-instate the projection source method
Re-instated the source method of provider and made the slices query method private to the provider implementation. The query method now also filters projection ids so that they are within the slice range.

EntityType remains distinct from a Topic name, but they can be the same in our example. Topics can be namespaced using a ':' delimiter so they should remain their own thing.