Benchmarking the projection
Principally, I've introduced a benchmark for the Streambed storage projection. I checked heap allocations and CPU and nothing untoward was found. My machine (M1Max) yields about 10 million events per second, and we're pretty much down to the time it takes for channels and streams to process.

The benchmark did highlight that both the source provider and handler traits should operate on a mutable self, and this is now consistent with their equivalents for the entity manager.