Re-open connection on transmission failures
The file descriptor exhaustion issue appeared to be caused by the producer failing when transmitting, and then not attempting to reconnect. Consequently, another subscription to the commit log was made and failed rapidly, and didn't provide enough time for the last one to clean up its resources. This rapid failure loop was not recovering either.

Another, albeit separate error, was that a pending handler failure wouldn't cause a new source to be made. This was an oversight.

I am no longer able to see the file descriptor exhaustion after many connects/re-starts of the JVM service having posted UDP observations.