Introduces and_then as a convenience
Given the widespread use case for `.and(then(`, we can provide a convenience combinator named `.and_then`. Rust developers are familiar with `.and_then` as a combinator and its semantics here are the same as in other parts of the ecosystem (as it is for `.and(then(`).

We could introduce other commonly used combinators if required in future. I'm unsure if we need to though.