Handle more gRPC error conditions (#102)
* Handle more gRPC error conditions

This commit improves the gRPC error when consuming streams from a gRPC endpoint. Unfortunately, Tonic doesn't allow us to test for a connection being reset specifically; it just returns an error. I think this is a feature that could be improved with Tonic. If we could test for connections being reset then we could offer more logging for

* Provides a convenience for marshalling declarations (#103)