Simplify with `then_reply`
`.and_then_reply` has been simplified with a `then_reply` which will only be called when the previous result was successful, and only the state is available to make a decision with. The closure is deliberately not async as it commonly isn't require for this use case.

More complex reply scenarios can still be achieved by using the all-flexible `.and_then` combinator.