Refactor "emit" to "persist" (#119)
Refactors "emit" as "persist" and attempts to convey the intention of associated operations that persistence is required.

As a consequence, the commit log marshaller is refactored to remove the optionality of some return types to support that all events must be persisted. Along the way, I also refactored some options into a result with increased error reporting and logging.