Entity manager and other conveniences (#121)
* Improve the API ergonomics

Improves the API ergonomics by providing a convenience function for running an entity manager and projections etc. The function returns both a task to spawn and a channel for sending commands.

Along the way, I also removed the `NonZeroUsize` type as a parameter as it is ugly, and not intended for APIs.

* Returns a filter of consumers

Replaces having to create a watch, and a convenience for providing an initial set of filters.

* Return envelopes and filters

For convenience, returns an envelope sender and a filter receiver from the task.